rm stats function and use lodahs instead
